move remote config to appsec folder
