+++ packages/dd-trace/src/appsec/remote_config/capabilities.js
