+++ packages/dd-trace/src/proxy.js
@@ -10,8 +10,8 @@ const { setStartupLogPluginManager } = require('./startup-log')
 const telemetry = require('./telemetry')
 const PluginManager = require('./plugin_manager')
 const { sendGitMetadata } = require('./ci-visibility/exporters/git/git_metadata')
-const RemoteConfigManager = require('./remote_config')
-const RemoteConfigCapabilities = require('./remote_config/capabilities')
+const RemoteConfigManager = require('./appsec/remote_config')
+const RemoteConfigCapabilities = require('./appsec/remote_config/capabilities')
 
 const gitMetadataUploadFinishCh = channel('ci:git-metadata-upload:finish')
