+++ packages/dd-trace/test/appsec/remote_config/scheduler.spec.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const Scheduler = require('../../src/remote_config/scheduler')
+const Scheduler = require('../../../src/appsec/remote_config/scheduler')
 
 describe('Scheduler', () => {
   let clock
